+++ docs/changelog.md
@@ -8,6 +8,12 @@
 
 See the [GitHub "Features" project](https://github.com/users/peter88213/projects/1).
 
+### v4.43.0
+
+- A file can be linked to each character, location, and item. 
+
+Based on PyWriter v12.15.3
+
 ### v4.42.4
 
 - Do not colorize scenes in "workphase" color mode, if the workphase is undefined.